+++ pdf_game/mod/pages.py
@@ -35,7 +35,7 @@ def render_victory(pdf, game_state, links_to_credits):
         pdf.image('assets/yellow-star.png', x=x, y=y)
     bitfont_render(pdf, 'VICTORY !', 80, 48, Justify.CENTER, size=24)
     bitfont_render(pdf, f'Secrets found: {len(game_state.secrets_found)}/4', 80, 80, Justify.CENTER)
-    query_param = urlsafe_b64encode(','.join(secrets_found).encode()).decode()
+    query_param = urlsafe_b64encode(','.join(game_state.secrets_found).encode()).decode()
     url = f'https://chezsoi.org/lucas/undying-dusk/hall-of-fame?v={__version__}&gs={query_param}'
     with bitfont_color_red():
         bitfont_render(pdf, 'Online hall of fame', 80, 90, Justify.CENTER, url=url)